+++ breeze-core/src/main/kotlin/domain/data/DataBuilder.kt
@@ -3,7 +3,12 @@
  * Breeze is blowing...
  ******************************************************************************/
 
-package com.windea.breezeframework.core.domain
+/*******************************************************************************
+ * Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+ * Breeze is blowing...
+ ******************************************************************************/
+
+package com.windea.breezeframework.core.domain.data
 
 /**
  * 数据构建器。