+++ breeze-core/src/main/kotlin/domain/data/Identifiable.kt
@@ -3,9 +3,14 @@
  * Breeze is blowing...
  ******************************************************************************/
 
+/*******************************************************************************
+ * Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+ * Breeze is blowing...
+ ******************************************************************************/
+
 @file:Suppress("UNCHECKED_CAST")
 
-package com.windea.breezeframework.core.domain
+package com.windea.breezeframework.core.domain.data
 
 import java.io.*