+++ breeze-http/src/main/kotlin/HttpConfig.kt
@@ -5,7 +5,7 @@
 
 package com.windea.breezeframework.http
 
-import com.windea.breezeframework.core.domain.*
+import com.windea.breezeframework.core.domain.data.*
 import java.net.*
 import java.net.http.*
 import java.time.*