+++ breeze-mapper/src/main/kotlin/impl/JsonMapper.kt
@@ -20,7 +20,7 @@ unmap
 
 */
 
-import com.windea.breezeframework.core.domain.*
+import com.windea.breezeframework.core.domain.data.*
 import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.mapper.*
 import java.lang.reflect.*