+++ breeze-mapper/src/main/kotlin/impl/XmlMapper.kt
@@ -5,11 +5,9 @@
 
 package com.windea.breezeframework.mapper.impl
 
-import com.windea.breezeframework.core.domain.*
-import com.windea.breezeframework.core.extensions.*
+import com.windea.breezeframework.core.domain.data.*
 import com.windea.breezeframework.mapper.*
 import java.lang.reflect.*
-import java.util.jar.*
 
 //TODO